requirements.txt
Flask_Cors==3.0.8
smpplib==2.0.1
redis==3.2.1
Flask==1.0.2
mysql_connector_repackaged==0.3.1

